+++ R/run_plink.R
@@ -75,7 +75,13 @@ run_plink <- function(
   error_line_index <- stringr::str_which(text, "Error:")
   if (length(error_line_index) != 0) {
     error_text <- text[seq(from = error_line_index, to = length(text))]
-    stop(paste0(error_text, collapse = " \n"))
+    stop(
+      paste0(error_text, collapse = " \n"),
+      " \n",
+      "Called PLINK with commands: \n",
+      normalizePath(plink_exe_path), " ", paste0(args, collapse = " "), "\n",
+      "Tip: you should be able to copy-paste this"
+    )
   }
   warnings <- stringr::str_subset(text, "Warning")
   if (length(warnings) != 0) warning(warnings)