+++ tests/testthat/test-create_assoc_args.R
@@ -20,6 +20,26 @@ test_that("minimal use", {
   )
 })
 
+test_that("v1.7", {
+  assoc_params <- create_test_assoc_params()
+  created <- create_assoc_args(
+    assoc_params = assoc_params,
+    plink_options = create_plink_v1_7_options()
+  )
+  expected <- c(
+    "--map", paste0(assoc_params$base_input_filename, ".map"),
+    "--ped", paste0(assoc_params$base_input_filename, ".ped"),
+    "--assoc",
+    "--maf", assoc_params$maf,
+    "--ci", assoc_params$confidence_interval,
+    "--out", assoc_params$base_output_filename
+  )
+  expect_equal(created, expected)
+
+  # PLINK v1.7 does not support this
+  expect_false("--allow-extra-chr" %in% created)
+})
+
 test_that("v1.9", {
   assoc_params <- create_test_assoc_params()
   created <- create_assoc_args(
@@ -29,6 +49,7 @@ test_that("v1.9", {
   expected <- c(
     "--map", paste0(assoc_params$base_input_filename, ".map"),
     "--ped", paste0(assoc_params$base_input_filename, ".ped"),
+    "--assoc",
     "--allow-extra-chr",
     "--maf", assoc_params$maf,
     "--ci", assoc_params$confidence_interval,
@@ -45,5 +66,8 @@ test_that("v2.0", {
     plink_options = create_plink_v2_0_options()
   )
   # No idea yet
-  expect_equal(created, expected)
+  expect_true("--glm" %in% created)
+
+  # PLINK v2.0 does not support this
+  expect_false("--assoc" %in% created)
 })